requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cloudinary==1.44.1
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg-binary==3.3.2
PyJWT==2.10.1
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
gunicorn==22.0.0
